refactor(snapshot): decompose common helpers
