+++ components/src/dynamo/common/snapshot/lifecycle.py
@@ -1,0 +1,174 @@
+# SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+"""Checkpoint lifecycle helpers for Dynamo Snapshot."""
+
+import asyncio
+import logging
+import os
+from dataclasses import dataclass
+from pathlib import Path
+from typing import Any, Generic, TypeVar
+
+from dynamo.common.snapshot.constants import (
+    READY_FOR_SNAPSHOT_FILE,
+    RESTORE_COMPLETE_FILE,
+    SNAPSHOT_COMPLETE_FILE,
+    SNAPSHOT_CONTROL_DIR_ENV,
+)
+
+logger = logging.getLogger(__name__)
+EngineT = TypeVar("EngineT")
+
+# Poll interval for the snapshot-control directory. Checkpoint and restore
+# latencies are seconds, so 100ms is negligible overhead.
+SENTINEL_POLL_INTERVAL_SEC = 0.1
+
+
+def is_checkpoint_enabled() -> bool:
+    return bool(os.environ.get(SNAPSHOT_CONTROL_DIR_ENV))
+
+
+class CheckpointConfig:
+    """Parsed checkpoint configuration plus the sentinel-driven lifecycle."""
+
+    def __init__(self, control_dir: str):
+        self.control_dir = control_dir
+        self.ready_file = os.path.join(control_dir, READY_FOR_SNAPSHOT_FILE)
+
+    @classmethod
+    def from_env(cls) -> "CheckpointConfig | None":
+        control_dir = os.environ.get(SNAPSHOT_CONTROL_DIR_ENV)
+        if not control_dir:
+            return None
+
+        configure_checkpoint_transport_env()
+        return cls(control_dir=control_dir)
+
+    async def run_lifecycle(
+        self,
+        pause_controller: Any,
+        *pause_args: object,
+    ) -> bool:
+        logger.info("Pausing model")
+        await pause_controller.pause(*pause_args)
+
+        try:
+            with open(self.ready_file, "w", encoding="utf-8") as ready_file:
+                ready_file.write("ready")
+
+            logger.info(
+                "Ready for checkpoint. Polling for sentinel in %s "
+                "(snapshot-complete or restore-complete)",
+                self.control_dir,
+            )
+
+            event = await self._wait_for_sentinel()
+        finally:
+            self._cleanup_ready_and_sentinels()
+
+        if event == "restore":
+            logger.info("Restore sentinel detected")
+            logger.info("Resuming model after restore")
+            await pause_controller.resume()
+            pause_controller.mark_resumed()
+            return True
+
+        logger.info("Snapshot completion sentinel detected")
+        return False
+
+    async def _wait_for_sentinel(self) -> str:
+        snapshot_path = Path(self.control_dir) / SNAPSHOT_COMPLETE_FILE
+        restore_path = Path(self.control_dir) / RESTORE_COMPLETE_FILE
+        while True:
+            if snapshot_path.exists():
+                return "checkpoint"
+            if restore_path.exists():
+                return "restore"
+            await asyncio.sleep(SENTINEL_POLL_INTERVAL_SEC)
+
+    def _cleanup_ready_and_sentinels(self) -> None:
+        for name in (
+            READY_FOR_SNAPSHOT_FILE,
+            SNAPSHOT_COMPLETE_FILE,
+            RESTORE_COMPLETE_FILE,
+        ):
+            path = os.path.join(self.control_dir, name)
+            try:
+                os.unlink(path)
+            except FileNotFoundError:
+                pass
+            except OSError:
+                logger.exception("Failed to clean up %s at %s", name, path)
+
+
+def configure_checkpoint_transport_env() -> None:
+    nccl_cumem_enable = os.environ.get("NCCL_CUMEM_ENABLE")
+    if nccl_cumem_enable and nccl_cumem_enable != "0":
+        logger.warning(
+            "Overriding NCCL_CUMEM_ENABLE=%r with '0' for checkpoint mode "
+            "because cuda-checkpoint does not support cuMem-backed NCCL allocations",
+            nccl_cumem_enable,
+        )
+    os.environ["NCCL_CUMEM_ENABLE"] = "0"
+
+    nccl_p2p_disable = os.environ.get("NCCL_P2P_DISABLE")
+    if nccl_p2p_disable and nccl_p2p_disable != "0":
+        logger.warning(
+            "Overriding NCCL_P2P_DISABLE=%r with '0' for checkpoint mode "
+            "to keep NCCL on GPU P2P transport when topology allows it",
+            nccl_p2p_disable,
+        )
+    os.environ["NCCL_P2P_DISABLE"] = "0"
+
+    nccl_nvls_enable = os.environ.get("NCCL_NVLS_ENABLE")
+    if nccl_nvls_enable and nccl_nvls_enable != "0":
+        logger.warning(
+            "Overriding NCCL_NVLS_ENABLE=%r with '0' for checkpoint mode "
+            "to avoid NVLS and keep NCCL on the legacy P2P path",
+            nccl_nvls_enable,
+        )
+    os.environ["NCCL_NVLS_ENABLE"] = "0"
+
+    nccl_ib_disable = os.environ.get("NCCL_IB_DISABLE")
+    if nccl_ib_disable and nccl_ib_disable != "1":
+        logger.warning(
+            "Overriding NCCL_IB_DISABLE=%r with '1' for checkpoint mode "
+            "because CRIU and cuda-checkpoint cannot restore InfiniBand state",
+            nccl_ib_disable,
+        )
+    os.environ["NCCL_IB_DISABLE"] = "1"
+
+    nccl_ras_enable = os.environ.get("NCCL_RAS_ENABLE")
+    if nccl_ras_enable and nccl_ras_enable != "0":
+        logger.warning(
+            "Overriding NCCL_RAS_ENABLE=%r with '0' for checkpoint mode "
+            "because NCCL RAS background state is not part of the checkpoint contract",
+            nccl_ras_enable,
+        )
+    os.environ["NCCL_RAS_ENABLE"] = "0"
+
+    torch_nccl_monitoring = os.environ.get("TORCH_NCCL_ENABLE_MONITORING")
+    if torch_nccl_monitoring and torch_nccl_monitoring != "0":
+        logger.warning(
+            "Overriding TORCH_NCCL_ENABLE_MONITORING=%r with '0' for "
+            "checkpoint mode because ProcessGroupNCCL monitoring can "
+            "terminate restored processes",
+            torch_nccl_monitoring,
+        )
+    os.environ["TORCH_NCCL_ENABLE_MONITORING"] = "0"
+    os.environ.setdefault("TORCH_NCCL_DUMP_ON_TIMEOUT", "0")
+
+
+@dataclass
+class EngineSnapshotController(Generic[EngineT]):
+    engine: EngineT
+    pause_controller: Any
+    checkpoint_config: CheckpointConfig
+    pause_args: tuple[object, ...] = ()
+
+    async def wait_for_restore(self) -> bool:
+        return await self.checkpoint_config.run_lifecycle(
+            self.pause_controller,
+            *self.pause_args,
+        )
