+++ components/src/dynamo/sglang/snapshot.py
@@ -9,7 +9,7 @@ import time
 
 import sglang as sgl
 
-from dynamo.common.snapshot import CheckpointConfig, EngineSnapshotController
+from dynamo.common.snapshot.lifecycle import CheckpointConfig, EngineSnapshotController
 
 from .pause import SGLangEnginePauseController
