+++ components/src/dynamo/vllm/main.py
@@ -23,7 +23,8 @@ from vllm.v1.engine.async_llm import AsyncLLM
 from vllm.v1.metrics.prometheus import setup_multiprocess_prometheus
 
 from dynamo.common.config_dump import dump_config
-from dynamo.common.snapshot import fetch_model_in_subprocess, is_checkpoint_enabled
+from dynamo.common.snapshot.lifecycle import is_checkpoint_enabled
+from dynamo.common.snapshot.model_fetch import fetch_model_in_subprocess
 from dynamo.common.snapshot.restore_context import (
     parse_snapshot_restore_runtime_config,
     refresh_snapshot_restore_config,
